Move ecobenefit search to treemap app
The initial thought a long time ago was that tree benefits calcuations
would have their own url, so that when you do a search, you would get
back tree benefits. As we continued to develop searching api and it
turned out that we will want to return more than ecobenfits. We may
want species data, tree counts, etc. So, we don't want this whole
other endpoint just for searching. As we kept adding more stuff, we
realized this doesn't really belong here anymore, it's not just going
to be eco benefits. So, we are moving it to the regular app. It's
going to grow into the main search function.

However, it may be useful down the road to again factor out the
portion that calculates eco benefits and but that back in
ecobenefits.views or in a utility function library like
ecobenefits.utils.
